# contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Yega API
  version: 0.1.0
  description: API for the Yega platform.
  license:
    name: Proprietary
    url: https://yega.com.mx/legal/api-license
servers:
  - url: https://api.stg.yega.com.mx
    description: Staging
  - url: https://api.yega.com.mx
    description: Production
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: User login
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/AuthInvalidCredentials'
  /auth/register:
    post:
      summary: User registration
      operationId: authRegister
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/AuthEmailTaken'
...
  /orders:
    get:
      summary: Get user orders
      operationId: ordersListMine
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: A list of user orders
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: Total number of orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new order
      operationId: ordersCreate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/OrderInvalidState'
  /orders/{orderId}:
    get:
      summary: Get order details
      operationId: ordersGetById
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/OrderNotFound'
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{orderId}/status:
    patch:
      summary: Update order status
      operationId: ordersUpdateStatus
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    AuthInvalidCredentials:
      description: Invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            auth.invalid_credentials:
              value:
                error:
                  code: auth.invalid_credentials
                  http: 401
                  message: Invalid credentials
                  details:
                    attempts_remaining: 2
    AuthEmailTaken:
      description: Email already registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            auth.email_taken:
              value:
                error:
                  code: auth.email_taken
                  http: 400
                  message: Email already registered
                  details:
                    email: user@example.com
    OrderNotFound:
      description: Order not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            orders.not_found:
              value:
                error:
                  code: orders.not_found
                  http: 404
                  message: Order not found
                  details:
                    order_id: 8b7f3c2d-4e5f-6a7b-8c9d-0e1f2a3b4c5d
    OrderInvalidState:
      description: Invalid order state for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            orders.invalid_state:
              value:
                error:
                  code: orders.invalid_state
                  http: 409
                  message: Invalid order state for this operation
                  details:
                    current: delivered
  schemas:
    User:
      type: object
      required: [id, name, email, role]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        role:
          type: string
          enum: [client, store, courier, admin]
...
    Order:
      type: object
      required: [id, userId, items, status, total]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          $ref: '#/components/schemas/OrderStatus'
        total:
          type: number
          format: float
          minimum: 0
        createdAt:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [productId, quantity, price]
      additionalProperties: false
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
          minimum: 0
    OrderStatus:
      type: string
      enum: [pending, accepted, preparing, assigned, on_route, delivered, canceled]
      description: |
        Estados válidos de un pedido:
        - pending: Pedido creado por el cliente
        - accepted: Tienda acepta el pedido
        - preparing: Tienda está preparando el pedido
        - assigned: Pedido asignado a un repartidor
        - on_route: Repartidor en camino a entregar
        - delivered: Pedido entregado
        - canceled: Pedido cancelado
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
        role:
          type: string
          enum: [client, store, courier]
    CreateOrderRequest:
      type: object
      required: [storeId, items]
      additionalProperties: false
      properties:
        storeId:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            additionalProperties: false
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
    AuthResponse:
      type: object
      required: [token, user]
      additionalProperties: false
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, http, message]
          additionalProperties: false
          properties:
            code:
              type: string
            http:
              type: integer
            message:
              type: string
            details:
              type: object
              additionalProperties: true